Fix to SE HAL deinit invoked even when logical channel is open
if logical channel is already open and then
basic channel open is called (with wrong AID) which results
in open failure, then basic channel close is invoked.
But this results in SE-HAL deinit even though logical
channel is still open.

So, Channel count is decremented only if its already opened.

Bug: 80448246
Test: VTS
Change-Id: Ifed516fef5eebd52f66ce59d3dc0e3910f4a0e53